added write method for output

+++ src/IO/MyCompressorOutputStream.java
@@ -57,7 +57,7 @@ public class MyCompressorOutputStream extends OutputStream {
         for (int i=0 ; i<array.size(); i++){ //send array
             write(array.get(i).getKey());
             for (int j=pointerSize-1 ; j>=0 ; j--)
-                write(array.get(i).getValue() >> 8*j );
+                write((byte)(array.get(i).getValue() >> 8*j) );
         }
         out.flush();
         //Finish writing compressed array//
